+++ src/nets.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-net::net() : connected_pins_(vector<unsigned>()), all_pins_(nullptr) {}
+net::net() : connected_pins_(vector<size_t>()), all_pins_(nullptr) {}
 
 net::net(net&& net)
     : connected_pins_(std::move(net.connected_pins_)),
@@ -13,7 +13,7 @@ net::net(net&& net)
     net.all_pins_ = nullptr;
 }
 
-net::net(vector<unsigned>&& conn, const vector<pin>& all_pins_)
+net::net(vector<size_t>&& conn, const vector<pin>& all_pins_)
     : connected_pins_(std::move(conn)), all_pins_(&all_pins_) {}
 
 net& net::operator=(net&& net) {
@@ -23,11 +23,11 @@ net& net::operator=(net&& net) {
     return *this;
 }
 
-unsigned net::at(unsigned index) const {
+size_t net::at(size_t index) const {
     return this->connected_pins_.at(index);
 }
 
-static pair<unsigned, unsigned> get_center(const pin& pin) {
+static pair<size_t, size_t> get_center(const pin& pin) {
     if (!pin.area_nonzero()) {
         assert(pin.width() == 0);
         assert(pin.height() == 0);
@@ -36,16 +36,16 @@ static pair<unsigned, unsigned> get_center(const pin& pin) {
                      pin.y() + (pin.height() >> 1));
 }
 
-unsigned net::hpwl() const {
+size_t net::hpwl() const {
     const auto& pin_list = *all_pins_;
 
-    unsigned max_w, min_w, max_h, min_h;
+    size_t max_w, min_w, max_h, min_h;
     auto center = get_center(pin_list[connected_pins_[0]]);
 
     max_w = min_w = center.first;
     max_h = min_h = center.second;
 
-    unsigned i, x, y;
+    size_t i, x, y;
     bool halt;
     for (i = 1, halt = false; i < connected_pins_.size(); ++i) {
         const pin& pin = pin_list[connected_pins_[i]];
